order end dialog added

+++ app/src/main/java/uz/gita/saiga_driver/data/remote/request/order/EndOrderRequest.kt
@@ -1,8 +1,7 @@
 package uz.gita.saiga_driver.data.remote.request.order
 
 data class EndOrderRequest(
-
-    val OrderLengthOfWay: Int,
-    val orderId: Int,
-    val orderMoney: Int
+    val OrderLengthOfWay: Double,
+    val orderId: Long,
+    val orderMoney: Double
 )
